// todo-app-frontend/src/api/models/BoardMembershipUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Todo App API
 * API for managing users and their todos. Now with WebSockets!
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Role } from './Role';
import {
    RoleFromJSON,
    RoleFromJSONTyped,
    RoleToJSON,
    RoleToJSONTyped,
} from './Role';

/**
 * 
 * @export
 * @interface BoardMembershipUpdate
 */
export interface BoardMembershipUpdate {
    /**
     * 
     * @type {Role}
     * @memberof BoardMembershipUpdate
     */
    role: Role;
}



/**
 * Check if a given object implements the BoardMembershipUpdate interface.
 */
export function instanceOfBoardMembershipUpdate(value: object): value is BoardMembershipUpdate {
    if (!('role' in value) || value['role'] === undefined) return false;
    return true;
}

export function BoardMembershipUpdateFromJSON(json: any): BoardMembershipUpdate {
    return BoardMembershipUpdateFromJSONTyped(json, false);
}

export function BoardMembershipUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): BoardMembershipUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'role': RoleFromJSON(json['role']),
    };
}

export function BoardMembershipUpdateToJSON(json: any): BoardMembershipUpdate {
    return BoardMembershipUpdateToJSONTyped(json, false);
}

export function BoardMembershipUpdateToJSONTyped(value?: BoardMembershipUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'role': RoleToJSON(value['role']),
    };
}